Fixed a bug when loading cached signatures in ImageScanner plugin

+++ src/org/dyndns/doujindb/plug/impl/imagescanner/NaiveSimilarityFinder.java
@@ -237,7 +237,7 @@ public class NaiveSimilarityFinder
 						* (g1 - g2) + (b1 - b2) * (b1 - b2));
 				dist += tempDist;
 			}
-		return 100 - (dist * 100 / (width * height * 0xff));
+		return 100 - (dist * 100 / ((double)width * height * 0xff));
     }
 	
 	private double distance_percent(BufferedImage bi)
